==> notebook/1_fetch data.ipynb <==
import wandb 
import pandas as pd

train = pd.read_csv(r"C:\Users\ASUS\Downloads\Mlop-prj-main\Mlop-prj-main\house-prices-advanced-regression-techniques\train.csv")
train.head(5)

train.info()

# save vào raw_data.csv 
train.to_csv("raw_data.csv", index=False)
# load vào wandb
import os 
from dotenv import load_dotenv
load_dotenv()

WANDB_API_KEY = os.environ.get("WANDB_API_KEY")

wandb.login()

run = wandb.init(project="predict_house_price", save_code=True)

artifact = wandb.Artifact(
    name="raw_data_csv",       
    type="raw_data",           
    description="Housing price raw data"  
)

artifact.add_file("raw_data.csv")  

run.log_artifact(artifact)                      

run = wandb.init(project="predict_house_price", job_type="load_data")

artifact = run.use_artifact('raw_data_csv:latest', type='raw_data')

artifact_dir = artifact.download()

file_path = os.path.join(artifact_dir, "raw_data.csv")

train = pd.read_csv(file_path)

run.finish()
train.info()
