==> notebook/3_preprocessing.ipynb <==
from ydata_profiling import ProfileReport
import os
from sklearn.preprocessing import LabelEncoder
import wandb
import pandas as pd 
from sklearn.decomposition import PCA
from sklearn.preprocessing import StandardScaler
from sklearn.model_selection import train_test_split
import xgboost as xgb
from sklearn.metrics import mean_squared_error , r2_score
import math  
import numpy as np 



run = wandb.init(project="predict_house_price", job_type="load_data")

artifact = run.use_artifact('clean_data_csv:latest', type='processed_data')

artifact_dir = artifact.download()

file_path = os.path.join(artifact_dir, "clean_data.csv")

clean_data = pd.read_csv(file_path, keep_default_na=False)

run.finish()


clean_data.info()

drop_cols = [
    "Id", "LowQualFinSF", "PoolArea", "PoolQC", "MiscVal"
]
clean_data = clean_data.drop(columns=drop_cols, errors="ignore")


# Gom nhóm các giá trị hiếm thành "Other"

# 1. Exterior1st (nhóm < 10 mẫu)
count_exterior1 = clean_data['Exterior1st'].value_counts()
rare_exterior1 = count_exterior1[count_exterior1 < 10].index.tolist()
clean_data['Exterior1st'] = clean_data['Exterior1st'].replace(rare_exterior1, 'Other')

# 2. Exterior2nd (nhóm < 10 mẫu)
count_exterior2 = clean_data['Exterior2nd'].value_counts()
rare_exterior2 = count_exterior2[count_exterior2 < 10].index.tolist()
clean_data['Exterior2nd'] = clean_data['Exterior2nd'].replace(rare_exterior2, 'Other')

# 3. BsmtFinType2 (nhóm < 100 mẫu)
count_bsmt = clean_data['BsmtFinType2'].value_counts()
rare_bsmt = count_bsmt[count_bsmt < 100].index.tolist()
clean_data['BsmtFinType2'] = clean_data['BsmtFinType2'].replace(rare_bsmt, 'Other')

# 4. GarageType (nhóm < 30 mẫu)
count_garage = clean_data['GarageType'].value_counts()
rare_garage = count_garage[count_garage < 30].index.tolist()
clean_data['GarageType'] = clean_data['GarageType'].replace(rare_garage, 'Other')

# 5. Fence (nhóm < 60 mẫu)
count_fence = clean_data['Fence'].value_counts()
rare_fence = count_fence[count_fence < 60].index.tolist()
clean_data['Fence'] = clean_data['Fence'].replace(rare_fence, 'Other')



from sklearn.preprocessing import OrdinalEncoder

# 1. Encode ordinal features
ordinal_cols = ["ExterQual", "KitchenQual", "BsmtQual"]
quality_order = ["None", "Po", "Fa", "TA", "Gd", "Ex"]

encoder = OrdinalEncoder(categories=[quality_order] * len(ordinal_cols))
clean_data[ordinal_cols] = encoder.fit_transform(clean_data[ordinal_cols])

# 2. Encode nominal features (còn lại sau khi xử lý rare)
ordinal_encoded = set(ordinal_cols)
all_cat_cols = clean_data.select_dtypes(include="object").columns.tolist()
nominal_cols = [col for col in all_cat_cols if col not in ordinal_encoded]

# Chỉ giữ các cột có số nhóm hợp lý (≤ 20)
nominal_cols_to_encode = [col for col in nominal_cols if clean_data[col].nunique() <= 20]

# 3. One-hot encoding cho các cột này
clean_data = pd.get_dummies(clean_data, columns=nominal_cols_to_encode, drop_first=True)

# 4. Frequency encoding cho biến nhiều nhóm (ví dụ: Neighborhood)
if "Neighborhood" in clean_data.columns:
    freq_map = clean_data["Neighborhood"].value_counts(normalize=True)
    clean_data["Neighborhood_FE"] = clean_data["Neighborhood"].map(freq_map)
    clean_data.drop(columns=["Neighborhood"], inplace=True)


clean_data.head()

run = wandb.init(project="predict_house_price", job_type="upload_preprocessing")

os.makedirs("preprocessed_data", exist_ok=True)

clean_data.to_csv("preprocessed_data/preprocessed_data.csv", index=False)

artifact = wandb.Artifact(
    name="preprocessed_data",       # tên artifact Kun muốn
    type="processed_data",           # kiểu dữ liệu
    description="Preprocessed dataset before training models"
)

artifact.add_file("preprocessed_data/preprocessed_data.csv")

run.log_artifact(artifact)

run.finish()

#ProfileReport(clean_data, title='kun', explorative=True)

# 2. Tách X và y
X = clean_data.drop(columns=["SalePrice"])
y = np.log1p(clean_data["SalePrice"])

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

run = wandb.init(project="predict_house_price", job_type='upload_pca_data')

output_dir = "pca_data"
os.makedirs(output_dir, exist_ok=True)


X_train.to_csv(os.path.join(output_dir, "X_train.csv"), index=True)
X_test.to_csv(os.path.join(output_dir, "X_test.csv"), index=True)
y_train.to_csv(os.path.join(output_dir, "y_train.csv"), index=True)
y_test.to_csv(os.path.join(output_dir, "y_test.csv"), index=True)

# Tạo artifact chung
artifact = wandb.Artifact(
    name="pca_X_y_train_test",
    type="processed_data",
    description="Dữ liệu X và y đã giảm chiều bằng PCA cho tập train/test"
)

# Add file vào artifact
artifact.add_file(os.path.join(output_dir, "X_train.csv"))
artifact.add_file(os.path.join(output_dir, "X_test.csv"))
artifact.add_file(os.path.join(output_dir, "y_train.csv"))
artifact.add_file(os.path.join(output_dir, "y_test.csv"))

# Upload lên wandb
run.log_artifact(artifact)

run.finish()
